--- frontend/src/pages/Orders.tsx
@@ -1,13 +1,0 @@
-import React from "react";
-
-const Orders = () => {
-    return (
-        <>
-            <div>Orders</div>
-            {/* TODO: table of orders */}
-            {/* TODO: export to csv */}
-        </>
-    );
-};
-
-export default Orders;
